refactor:task-003:フォルダ構成を変更
