+++ lib/core/models/screen.dart
