+++ lib/core/utils/get_category.dart
@@ -1,5 +1,5 @@
-import 'package:kakeibo/utils/get_expenditure_category.dart';
-import 'package:kakeibo/utils/get_income_category.dart';
+import 'package:kakeibo/core/utils/get_expenditure_category.dart';
+import 'package:kakeibo/core/utils/get_income_category.dart';
 
 String getCategory(int categoryType, int category) {
   switch (categoryType) {
