+++ lib/core/utils/get_expenditure_category.dart
