+++ lib/core/utils/get_income_category.dart
