+++ lib/features/history/widget/search_card.dart
