+++ lib/features/home/viewmodel/home_view_model.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:kakeibo/ui/components/views/home_screen.dart';
+import 'package:kakeibo/features/home/view/home_screen.dart';
 
 class HomeViewModel extends StatefulWidget {
   const HomeViewModel({super.key});
