+++ lib/main.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:kakeibo/ui/my_app.dart';
+import 'package:kakeibo/my_app.dart';
 
 void main() {
   runApp(const MyApp());
